requirements.txt
Django==3.1.2
djangorestframework==3.13.1
drf-yasg==1.20.0
elasticsearch==8.6.1
Faker==14.2.0
psycopg2-binary==2.9.3
pytest==7.1.3
pytest-django==4.5.2
pytest-env==0.6.2
pytest-forked==1.4.0
pytest-mock==3.8.2
python-decouple==3.6
